+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 ### Changed
 - Adopt semver via axion-release plugin
 - Target MC version 1.15
--
+
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).